clean up normalize function, split it into files
