+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/compiler/normalizer/package.scala
@@ -1,0 +1,3 @@
+package coop.rchain.rholang.interpreter.compiler
+
+package object normalizer {}
