+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/compiler/normalizer/procNormalizeMatcher/PEvalNormalizer.scala
@@ -1,0 +1,23 @@
+package coop.rchain.rholang.interpreter.compiler.normalizer
+
+import cats.syntax.all._
+import cats.effect.Sync
+import coop.rchain.models.Par
+import coop.rchain.models.rholang.implicits._
+import coop.rchain.rholang.interpreter.compiler.{NameVisitInputs, ProcVisitInputs, ProcVisitOutputs}
+import coop.rchain.rholang.ast.rholang_mercury.Absyn.PEval
+
+object PEvalNormalizer {
+  def normalize[F[_]: Sync](p: PEval, input: ProcVisitInputs)(
+      implicit env: Map[String, Par]
+  ): F[ProcVisitOutputs] =
+    NameNormalizeMatcher
+      .normalizeMatch[F](p.name_, NameVisitInputs(input.env, input.knownFree))
+      .map(
+        nameMatchResult =>
+          ProcVisitOutputs(
+            input.par ++ nameMatchResult.chan,
+            nameMatchResult.knownFree
+          )
+      )
+}
